# api/auth/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fanatikos Auth API",
    "version": "1.0.0",
    "description": "API documentation for authentication endpoints."
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "/health",
        "description": "Check API health.",
        "responses": {
          "200": {
            "description": "API is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "ok" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "/login",
        "description": "Authenticate user and return JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": { "type": "string" }
                      }
                    },
                    "message": { "type": "string", "example": "Login success" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "/register",
        "description": "Register a new user and return JWT token.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "username": { "type": "string" },
                            "email": { "type": "string", "format": "email" }
                          }
                        },
                        "access_token": { "type": "string" }
                      }
                    },
                    "message": { "type": "string", "example": "User registered successfully" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Username already exists"
          }
        }
      }
    },
    "/me": {
      "get": {
        "summary": "/me",
        "description": "Get current user info (JWT required).",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Current user info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "type": "string", "description": "user id" },
                    "info": { "type": "object", "description": "JWT info" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/validate": {
      "get": {
        "summary": "/validate",
        "description": "Validate JWT token.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Token is valid" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}